Additional parallel work and begin adding timing output

+++ create-st-geo-parallel.py
@@ -54,10 +54,14 @@ def download_geo(st):
 
     #print(STURL)
 
-    try:
-        urllib.request.urlretrieve(STURL,GEOPATH + FILENAME)
-    except urllib.error.HTTPError as e:
-        print(FILENAME + ': ' + e)
+    if not os.path.isfile(GEOPATH + FILENAME):
+        try:
+            urllib.request.urlretrieve(STURL,GEOPATH + FILENAME)
+            print("Downloaded ",FILENAME)
+        except urllib.error.HTTPError as e:
+            print(FILENAME + ': ' + e)
+    else:
+        print("Skipping ",FILENAME)
 
 ## Download state-level csv files into geo directory
 print("Downloads beginning.")
